separate waitForDelay
